+++ src/pdoc-commons/services/pdoc-data.service.ts
@@ -72,7 +72,7 @@ export class PDocDataService extends CommonDocDataService<PDocRecord, PDocSearch
                 profiles_ss: {
                     'in': [record.profiles]
                 },
-                type_txt: {
+                type_ss: {
                     'in': [record.type.toLowerCase()]
                 }
             }