+++ src/pdoc-commons/services/staticpages-data.service.ts
@@ -61,7 +61,7 @@ export class StaticPagesDataService extends CommonDocDataService<PDocRecord, PDo
                 name_s: {
                     'in': [record.name]
                 },
-                type_txt: {
+                type_ss: {
                     'in': [record.type.toLowerCase()]
                 }
             }